# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Banque Laravel",
        "description": "API REST complète pour la gestion bancaire - Comptes, Clients, Authentification",
        "contact": {
            "email": "contact@banque.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://moustapha-seck.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    " Authentification"
                ],
                "summary": " Connexion utilisateur (Admin ou Client)",
                "description": "Authentification d'un administrateur ou d'un client avec génération de token JWT. Les admins n'ont pas besoin de code, les clients oui pour la première connexion.",
                "operationId": "5db09a35f965c6ca7529ed3c3cae8e21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin1@banque.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe",
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": " Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "e8061d9e-1385-4fb9-849c-0bd2945fc49c"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "admin1@banque.com"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Administrateur 1",
                                                            "nullable": true
                                                        },
                                                        "titulaire": {
                                                            "type": "string",
                                                            "example": "Fatou Sow",
                                                            "nullable": true
                                                        },
                                                        "telephone": {
                                                            "type": "string",
                                                            "example": "+221782345678",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": " Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "INVALID_CREDENTIALS"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Email ou mot de passe incorrect"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": " Code de vérification requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "CODE_REQUIRED"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Code de vérification requis pour la première connexion"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    " Authentification"
                ],
                "summary": " Rafraîchir le token d'accès",
                "description": "Génère un nouveau token d'accès et révoque l'ancien",
                "operationId": "05bd0482845348cafa4da9e4fb7d0ffb",
                "responses": {
                    "200": {
                        "description": " Token rafraîchi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token rafraîchi"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": " Token invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "UNAUTHENTICATED"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token invalide"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    " Authentification"
                ],
                "summary": " Déconnexion",
                "description": "Révoque le token d'accès actuel",
                "operationId": "8f1a443d39c3bf383f851fbb51253c64",
                "responses": {
                    "200": {
                        "description": " Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token JWT obtenu via POST /api/v1/auth/login"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}